+++ internal/accounts/server_type_test.go
@@ -25,9 +25,30 @@ func (s *AccountTypeSuite) TestDescription() {
 }
 
 func (s *AccountTypeSuite) TestAccountTypeFromURL() {
-	s.Equal(ServerTypeShinyappsIO, serverTypeFromURL("https://api.shinyapps.io"))
-	s.Equal(ServerTypeShinyappsIO, serverTypeFromURL("https://api.staging.shinyapps.io"))
-	s.Equal(ServerTypeCloud, serverTypeFromURL("https://api.posit.cloud"))
-	s.Equal(ServerTypeCloud, serverTypeFromURL("https://api.rstudio.cloud"))
-	s.Equal(ServerTypeConnect, serverTypeFromURL("https://example.com"))
+	for _, test := range []struct {
+		url        string
+		serverType ServerType
+	}{
+		{"https://api.shinyapps.io", ServerTypeShinyappsIO},
+		{"https://api.shinyapps.io/connect/#/content", ServerTypeShinyappsIO},
+		{"https://api.staging.shinyapps.io", ServerTypeShinyappsIO},
+		{"https://api.staging.shinyapps.io/connect/#/content", ServerTypeShinyappsIO},
+		{"https://api.posit.cloud", ServerTypeCloud},
+		{"https://api.posit.cloud/connect/#/content", ServerTypeCloud},
+		{"https://api.rstudio.cloud", ServerTypeCloud},
+		{"https://api.rstudio.cloud/connect/#/content", ServerTypeCloud},
+		{"https://example.com", ServerTypeConnect},
+		{"https://example.com/connect/#/content", ServerTypeConnect},
+		{"https://example.snowflakecomputing.app", ServerTypeSnowflake},
+		{"https://example.snowflakecomputing.app/connect/#/content", ServerTypeSnowflake},
+		{"https://example.privatelink.snowflake.app", ServerTypeSnowflake},
+		{"https://example.privatelink.snowflake.app/connect/#/content", ServerTypeSnowflake},
+	} {
+		serverType, err := ServerTypeFromURL(test.url)
+		s.Nil(err)
+		s.Equal(test.serverType, serverType)
+	}
+
+	_, err := ServerTypeFromURL(":bad")
+	s.NotNil(err)
 }