+++ extensions/vscode/CHANGELOG.md
@@ -6,6 +6,12 @@ file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added support for deploying to Connect in Snowflake using key-pair auth. (#2632)
+
 ## [1.14.0]
 
 ### Added